% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5_compare3D_brute.R,
%   R/step5_compare3D_brute_tmp.R
\name{calculateCCFmaxNew}
\alias{calculateCCFmaxNew}
\title{Compute maximum correlation for two images, taking into account a sequence of
rotation angles and all integer-valued translations}
\usage{
calculateCCFmaxNew(image1, image2, pad = TRUE)

calculateCCFmaxNew(image1, image2, pad = TRUE)
}
\arguments{
\item{image1}{first image matrix}

\item{image2}{second image matrix. The two images do not need to be the same
size.}

\item{pad}{logical value for whether the smaller image should be zero-padded.
The default is TRUE. Only set this to FALSE if the images are of the same
size and are square.}

\item{I1}{image that is transformed}

\item{I2}{image to align to}
}
\value{
A list with four items: 1) maximum correlation taking into account
  many possible rotations and translations of the second image, 2) the
  corresponding dx, the horizontal translation, 3) dy, the vertical
  translation, and 4) theta, the corresponding rotation angle.

transformation vector \code{p} (theta, dx, dy), number of iterations
 before convergence, final correlation (only breechface area), MSE (only
 breechface area), overlapping number of pixels in breechface area
}
\description{
Given two processed images, the images are first resized to 1/4 of their
original dimension. The smaller image is zero-padded to make the two images
the same size. Then the second image is rotated through all rotation angles
2.5 degrees apart, and for each rotation we compute the maximum correlation
between the two images, considering all integer-valued translations in the x-
and y-dimensions. We then choose the rotation angle that results in the
largest correlation. We then rotate 0.5 degrees around this best rotation
angle, and again compute the maximum correlation. This serves as a similarity
metric for the two input images.

Runs the Lucas-Kanade algorithm to find best alignment parameters and returns
correlation and mean-squared error
}
\examples{
\dontrun{
calculateCCFmax(processedExample, processedExample2)
}

\dontrun{
# first run this:
}

}
