% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5_compare3D_brute.R
\name{calculateCCFmaxNorm}
\alias{calculateCCFmaxNorm}
\title{Normalize image by mean and sum of squares before computing maximum
cross-correlation}
\usage{
calculateCCFmaxNorm(image1, image2)
}
\arguments{
\item{image1}{first image matrix}

\item{image2}{second image matrix. The two images should be the same size.}
}
\value{
A list with four items: 1) maximum correlation taking into account
  many possible rotations and translations of the second image, 2) the
  corresponding dx, the horizontal translation, 3) dy, the vertical
  translation, and 4) theta, the corresponding rotation angle.
}
\description{
Does the same thing as `calculateCCFmaxNew()`, except images are normalized
by the square-root of the sum of squares. With both mean and sum of squares
normalized, maximizing cross-correlation is equivalent to minimizing
mean-squared error.
}
\examples{
\dontrun{
calculateCCFmax(processedExample, processedExample2)
}

}
