% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_selectBF_3D.R
\name{findPlaneRansac}
\alias{findPlaneRansac}
\title{Finds plane of breechface marks using RANSAC}
\usage{
findPlaneRansac(inputDepths)
}
\arguments{
\item{inputDepths}{matrix of input depths in microns.}
}
\value{
List object containing fitted plane and selected breechface marks.
}
\description{
Given input depths (in microns), find best-fitting plane using RANSAC. This
should be the plane that the breechface marks are on.
}
\examples{
\dontrun{
    set.seed(0)
    testImage <- findPlaneRansac(surfaceMat)
}
}
