% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_selectBF_3D.R
\name{levelBF3D}
\alias{levelBF3D}
\title{Levels image}
\usage{
levelBF3D(ransacOutput)
}
\arguments{
\item{ransacOutput}{list output from \code{findPlaneRansac}}
}
\value{
Leveled image
}
\description{
Levels image using plane fitted in \code{findPlaneRansac}.
}
\examples{
\dontrun{
    testImage <- levelBF3D(testImage)
}
}
