% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5_compare3D.R
\name{padImages}
\alias{padImages}
\title{Zero pad pairs of images}
\usage{
padImages(I1, I2)
}
\arguments{
\item{I1}{first image}

\item{I2}{second image}
}
\value{
list with two images
}
\description{
Given a pair of images, zero-pad smaller one so that they are the same size
}
\examples{
\dontrun{
# first run this:
}

}
