% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_removeCircular3D.R
\name{removeCircular3D}
\alias{removeCircular3D}
\title{Remove circular symmetry of an image}
\usage{
removeCircular3D(inputImage)
}
\arguments{
\item{inputImage}{701 x 701 matrix}
}
\value{
A 701 x 701 image matrix.
}
\description{
Given a 701 x 701 image matrix, fits a circularly symmetric model on the
image and returns the residuals. Areas that are not part of the breechface
impression are set to NA.
}
\examples{
\dontrun{
# first run this:
centeredExample <- centerBFprimer(FPexample, primerExample)
leveledExample <- levelBF(centeredExample$centeredBF)
# then
removedExample <- removeCircular(leveledExample)
}

}
