% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_standardize3D.R
\name{standardize3D}
\alias{standardize3D}
\title{Standardize 3D images}
\usage{
standardize3D(surfaceMat, lateralResMicrons)
}
\arguments{
\item{surfaceMat}{matrix of depth values in microns}

\item{lateralResMicrons}{lateral resolution of image}
}
\value{
Standardized image for further pre-processing.
}
\description{
Standardize topographies so that they have the same lateral resolution (6.25
microns), and crop/pad images such that they are 701x701 pixel images (NOTE:
first find plane and level, and then only standardize before further
pre-processing)
}
\examples{
\dontrun{
    testImage <- standardize3D(testImage, lateralResMicrons)
}
}
