% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess3D.R
\name{allPreprocess3D}
\alias{allPreprocess3D}
\title{Perform all pre-processing steps}
\usage{
allPreprocess3D(fileName)
}
\arguments{
\item{fileName}{location of image file. This function has been tested on
\code{png} files.}
}
\value{
An image matrix, cropped to the smallest rectangle containing the
  valid breechface area. Non-breechface pixels are set to 0. After all
  pre-processing, possible pixel values are -510 to 510.
}
\description{
Images must be of the standard format as provided in the NIST Ballistics and
Research Toolmark Database. The dimensions are 1944 x 2592 pixels, and images
are on a 255-grayscale. Here we use ring light images. This function performs
all the pre-processing steps, consisting of \itemize{ \item{Automatically
selecting the breechface marks} \item{Leveling the image} \item{Removing
circular symmetry} \item{Removing outliers and filtering.} }
}
\examples{
\dontrun{
}

}
