% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{standardize3D}
\alias{standardize3D}
\title{Standardize 3D images}
\usage{
standardize3D(surfaceMat, lateralResMicrons)
}
\arguments{
\item{surfaceMat}{matrix of depth values in microns}

\item{lateralResMicrons}{lateral resolution of image}
}
\value{
Standardized image for further pre-processing.
}
\description{
Standardize topographies so that they have the same lateral resolution (6.25
microns), and center and crop images (NOTE: first find plane and then
standardize before further pre-processing) (TODO center and crop)
}
\examples{
\dontrun{
    testImage <- standardize3D(testImage, lateralResMicrons)
}
}
